+++ packages/apptive_grid_core/example/lib/main.dart
@@ -7,8 +7,9 @@ void main() async {
     environment: ApptiveGridEnvironment.alpha,
     authenticationOptions: ApptiveGridAuthenticationOptions(
       autoAuthenticate: true,
+      authGroup: 'apptivegrid',
       redirectScheme: 'apptivegrid',
-      persistCredentials: true,
+      persistCredentials: false,
     ),
   );
   await enableWebAuth(options);