+++ packages/apptive_grid_core/lib/src/network/environment.dart
@@ -1,20 +1,17 @@
 /// Possible Stages/Environment ApptiveGrid can point to
 enum ApptiveGridEnvironment {
   /// Alpha Environment
-  alpha(url: 'https://alpha.apptivegrid.de', authRealm: 'apptivegrid-test'),
+  alpha(url: 'https://alpha.apptivegrid.de'),
 
   /// Beta Environment
-  beta(url: 'https://beta.apptivegrid.de', authRealm: 'apptivegrid-test'),
+  beta(url: 'https://beta.apptivegrid.de'),
 
   /// Production Environment
-  production(url: 'https://app.apptivegrid.de', authRealm: 'apptivegrid');
+  production(url: 'https://app.apptivegrid.de');
 
   /// Create a new ApptiveGridEnvironment
-  const ApptiveGridEnvironment({required this.url, required this.authRealm});
+  const ApptiveGridEnvironment({required this.url});
 
   /// The API url for the selected [ApptiveGridEnvironment]
   final String url;
-
-  ///The authentication realm that needs to be used for Authentication
-  final String authRealm;
 }