+++ packages/apptive_grid_core/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.13.1
-* Prase multiline text to single line for single line form fields
+* Parse multiline text to single line for single line form fields
 
 ## 0.13.0
 * Remove Deprecated Members